feat: eonline parser
I’m sorry, Adam, but for this one I couldn’t figure out how to parse
the content, so I was unable to also do the cleaning parts as well.

The date test failed, but the failure is do to a difference in
formatting (not sure how to correct this). It looks like it’s getting
the correct info, but there is a difference in the number of zeros.

Sorry, this one is a hot mess.