feat: accu extractor
This one might have been a bit too tough for me. Notes:
1) I couldn’t find a date_published in this article
2) The dek was nested within the content. I found a selector, but I
wasn’t sure how to return the content minus the dek, if that makes
sense.
3) I didn’t see an image, so I didn’t bother trying to enter a selector.
4) The initial complaint on Zen desk: element labeled "Figure 1" is
getting stripped out

Unfortunately the custom parser doesn’t solve this issue. I am guessing
I would need to do a transform… but I am not really clear on how to do
transforms. Might need to sit down with someone to get a tutorial on
this. :-)