fix: Initialize Content-Type as empty string if not present
This may be exceedingly rare, but if the response does not include a
Content-Type header, the parser will raise an exception. This fix ensures that
even if the header is not present, the value we check is still a string.

I discovered this issue while trying to parse my own site, and while I've fixed
the headers there, I left one bad URL for verification:
https://johnholdun.com/404