fix: arstechica.com
Pagination worked for full URLs, but not in the short URLs that Ars Technica puts in its RSS feeds. The
reason was that the parser takes the request URL into account when determining whether a link might be
to a subsequent page. That uses the requested URL, not the served URL.

I ended up implementing the ability to get the attribute with a JavaScript function instead of a simple
selector.